// src/widgets/edit-invoice/model/types.ts
export interface EditInvoiceFormValues {
  counterparty_id: string;
  service_name: string;
  service_code: string;
  unit_name: string;
  quantity: string;
  price: string;
  vat_rate: string;
  special_sales_book: string;
  inter_price_difference: string;
  correction_number: string;
  payment_document_number: string;
  payment_date: string;
}